a few tweaks

+++ planetiler-custommap/src/main/java/com/onthegomap/planetiler/custommap/configschema/AttributeDefinition.java
@@ -9,7 +9,7 @@ public record AttributeDefinition(
   @JsonProperty("tag_value") String tagValue,
   @JsonProperty("include_when") Map<String, Object> includeWhen,
   @JsonProperty("exclude_when") Map<String, Object> excludeWhen,
-  @JsonProperty("min_zoom") Byte minZoom,
-  @JsonProperty("min_zoom_by_value") Map<Object, Byte> minZoomByValue,
+  @JsonProperty("min_zoom") Integer minZoom,
+  @JsonProperty("min_zoom_by_value") Map<Object, Integer> minZoomByValue,
   @JsonProperty("min_tile_cover_size") Double minTileCoverSize
 ) {}
